Got rid of the warning for missing column indices FirstName and LastName in DataTable::printPedigreeTableAsTabDelimited() --RK
git-svn-id: https://svn.code.sf.net/p/madeline2/code/madeline2/trunk@53 36be61f0-9d1b-0410-b0b4-cea5cbe2d1c6

+++ src/DataTable.cpp
@@ -838,8 +838,8 @@ void DataTable::printPedigreeTableAsTabDelimited(std::string filename) const{
 	unsigned genderColumnIndex = getColumnOrdinal("Gender");
 	unsigned fatherColumnIndex = getColumnOrdinal("Father");
 	unsigned motherColumnIndex = getColumnOrdinal("Mother");
-	unsigned firstNameColumnIndex = (columnExists("FirstName")?getColumnOrdinal("FirstName"):COLUMN_IS_MISSING);
-	unsigned lastNameColumnIndex = (columnExists("LastName")?getColumnOrdinal("LastName"):COLUMN_IS_MISSING);
+	int firstNameColumnIndex = (columnExists("FirstName")?(int)getColumnOrdinal("FirstName"):COLUMN_IS_MISSING);
+	int lastNameColumnIndex = (columnExists("LastName")?(int)getColumnOrdinal("LastName"):COLUMN_IS_MISSING);
 	unsigned j;
 	
 	std::ofstream pedfile(filename.c_str());
